refactor: extract UnwrapClsDialog to own component
