+++ src/web/src/views/workspace/WSEditorCommandArgumentsContent.tsx
@@ -8,10 +8,7 @@ import {
   DialogContent,
   DialogTitle,
   LinearProgress,
-  styled,
   TextField,
-  Typography,
-  TypographyProps,
 } from "@mui/material";
 
 import { commandApi, errorHandlerApi } from "../../services";
@@ -21,13 +18,7 @@ import WSECArgumentSimilarPicker, { ArgSimilarTree, BuildArgSimilarTree } from "
 
 import ArgumentNavigation from "./ArgumentNavigation";
 import ArgumentDialog from "./ArgumentDialog";
-
-const ArgTypeTypography = styled(Typography)<TypographyProps>(({ theme }) => ({
-  color: theme.palette.primary.main,
-  fontFamily: "'Roboto Condensed', sans-serif",
-  fontSize: 16,
-  fontWeight: 700,
-}));
+import UnwrapClsDialog from "./UnwrapClsDialog";
 
 import { CardTitleTypography } from "./WSEditorTheme";
 
@@ -458,76 +449,6 @@ function FlattenDialog(props: {
   );
 }
 
-function UnwrapClsDialog(props: {
-  commandUrl: string;
-  arg: CMDArg;
-  open: boolean;
-  onClose: (unwrapped: boolean) => Promise<void>;
-}) {
-  const [updating, setUpdating] = useState<boolean>(false);
-  const [invalidText, setInvalidText] = useState<string | undefined>(undefined);
-
-  const handleClose = () => {
-    setInvalidText(undefined);
-    props.onClose(false);
-  };
-
-  const handleUnwrap = async () => {
-    setUpdating(true);
-
-    let argVar = props.arg.var;
-    if (props.arg.type.startsWith("array")) {
-      if ((props.arg as CMDArrayArg).item?.type.startsWith("@")) {
-        argVar += "[]";
-      }
-    } else if (props.arg.type.startsWith("dict")) {
-      if ((props.arg as CMDDictArg).item?.type.startsWith("@")) {
-        argVar += "{}";
-      }
-    }
-
-    const flattenUrl = `${props.commandUrl}/Arguments/${argVar}/UnwrapClass`;
-
-    try {
-      await commandApi.unwrapClassArgument(flattenUrl);
-      setUpdating(false);
-      await props.onClose(true);
-    } catch (err: any) {
-      console.error(err);
-      setInvalidText(errorHandlerApi.getErrorMessage(err));
-      setUpdating(false);
-    }
-  };
-
-  return (
-    <Dialog disableEscapeKeyDown open={props.open} sx={{ "& .MuiDialog-paper": { width: "80%" } }}>
-      <DialogTitle>Unwrap Class Type </DialogTitle>
-      <DialogContent dividers={true}>
-        {invalidText && (
-          <Alert variant="filled" severity="error">
-            {" "}
-            {invalidText}{" "}
-          </Alert>
-        )}
-        <ArgTypeTypography>{props.arg.type}</ArgTypeTypography>
-      </DialogContent>
-      <DialogActions>
-        {updating && (
-          <Box sx={{ width: "100%" }}>
-            <LinearProgress color="secondary" />
-          </Box>
-        )}
-        {!updating && (
-          <>
-            <Button onClick={handleClose}>Cancel</Button>
-            <Button onClick={handleUnwrap}>Unwrap Class</Button>
-          </>
-        )}
-      </DialogActions>
-    </Dialog>
-  );
-}
-
 type CMDArgHelp = {
   short: string;
   lines?: string[];
@@ -602,14 +523,8 @@ interface CMDDictArgBase extends CMDArgBase {
   item?: CMDArgBase;
   anyType: boolean;
 }
-interface CMDDictArg extends CMDDictArgBase, CMDArg {}
-
 interface CMDArrayArgBase extends CMDArgBase {
   item: CMDArgBase;
-}
-
-interface CMDArrayArg extends CMDArrayArgBase, CMDArg {
-  singularOptions?: string[];
 }
 
 type ClsArgDefinitionMap = {
